+++ CHANGELOG.md
@@ -11,4 +11,5 @@ There are next changes:
 - null for redefined constants was allowed [https://github.com/badoo/soft-mocks/pull/11](https://github.com/badoo/soft-mocks/pull/11), thanks [Alexey Manukhin](https://github.com/axxapy);
 - error "Fatal error: Couldn't find constant QA\SoftMocks::CLASS in /src/QA/SoftMocks.php on line 388" was fixed for old versions hhvm [https://github.com/badoo/soft-mocks/pull/16](https://github.com/badoo/soft-mocks/pull/16), thanks [Mougrim](https://github.com/mougrim);
 - warning "PHP Warning:  array_key_exists() expects exactly 2 parameters" was fixed [https://github.com/badoo/soft-mocks/pull/14](https://github.com/badoo/soft-mocks/pull/14), thanks [Mougrim](https://github.com/mougrim);
+- handle phpunit wrapped exceptions (PHPUnit_Framework_ExceptionWrapper, \PHPUnit\Framework\ExceptionWrapper);
 - unit tests was added.